Solve problem with padding in PlayerControls
Kirigami.Page is always adding padding around the content items.  This
gives visual artifacts with the swipepage on the player controls.  Part
of the previous page was still visible because the width of the
swipepage entries is less than the width of the full Kirigami.Page that
they are part of.