Fix app-crashing binding loop in FeedListPage
The binding loop occured whenever the vertical scrollbar appears or
disappears.  The GUI then gets stuck between solutions: one with a
certain amount of columns and another with one column less or more.  It
then keeps alternating between these two until the app freezes the
machine and eventually crashes.
The fix is to calculate the number of columns based on the total width
of the Page itself, rather than the width of the content.  By
subtracting enough space for the scrollbar, the cards on the page will
now rescale smoothly and properly.