Remove network connectivity check
Apparently in some cases the network manager is incorrectly signaling
that the system is offline while it's actually online. The current
implementation does not have a way to circumvent that incorrect
reporting. Also the error reporting made it seem that the actual
problem is a metered connection. Ripping out the connectivity check
solves these issues at the expense of connections timing out rather than
the connection not being set up in the first place when the system is
offline.

CCBUG: 475400