Speed up feed updates dramatically
Instead of always parsing the entire RSS/atom feed, we keep a hash of
the contents of the feed in the database.  If the hash of the new feed
retrieval matches the hash in the database, we know nothing has changed
and we can skip the detailed parsing of the feed altogether.