Fix 1px left and bottom border on mobile
In mobile mode, when the drawer is hidden, for some reason a 1px grey
border is drawn on the left of the application window. Change it to a
loader to avoid needlessly having it loaded, and remove the grey border.

There's also a 1 px window border on the bottom due to the navbar
thinking it's a header. Manually specify that it's in the footer.