Fix clickable links to feeds and entries
The current links were broken after the recent refactoring. The links
are opened through functions on the current page, but those functions
need to destroy the page they are part of. It seems like the objects
required to create and show the new pages would go out of scope before
the new pages were created properly.