Check for failed libvlc_new and abort the app properly if needed
This is a common error when a distros libvlc package does not contain
everything that's required to run Kasts.  For now, all we can do is warn
the user.  The backend will need a thorough refactoring to be able to
fall back to another backend, which would be the desired behaviour.